Made tower control function
Moved tower control code out of main into its own function, also fixed a
bug where only one tower in a room did anything
